Create early parse exceptions for non-parsed AST nodes
Create exceptions for AST nodes that are marked to never be valid in an executable AST after the last AST rewrite step. At this point we for sure know that these nodes are invalid in the AST and that a compile error should be created for them. Note that this does not catch wrong usage of CLabel and CEntry, since these are valid within certain functions.